core: tracking HTTP responses codes over inet sockets
This patch adds the new HTTP responses codes metric.

The eBPF program searches into the buffer the string `HTTP/1.1 XXX`,
where XXX is the 3 digits HTTP response code. The statistics are exposed
to Weave Scope using the report interface.

`tcp_sendmsg()` invokes `copy_from_iter()` multiple times to copy data from the iovec.
We inspect the data (`addr`) passed to `copy_from_iter()` only for its first invocation.
This limitation does not impact significantly the metric because is very unlikely
that a response code is splitted in multiple iovec or is not at the beginning of
the first iovec's element.

Codes tracked:
      - 100: Continue
      - 200: OK
      - 201: Created
      - 202: Accepted
      - 204: No Content
      - 308: Permanent Redirect Redirect
      - 400: Bad Request
      - 401: Unauthorized
      - 403: Forbidden
      - 404: Not Found
      - 408: Request Timeout
      - 451: Unavailable For Legal Reasons
      - 500: Internal Server Error
      - 501: Not Implemented
      - 502: Bad Gateway
      - 503: Service Unavailable